Reuse SparseMLTrainer for token classification
